requirements.txt
sentence-transformers
faiss-cpu
pandas
gradio
numpy